requirement.txt
amqp==2.1.4
billiard==3.5.0.2
celery==4.0.2
Django==1.11
django-background-tasks==1.1.10
django-compat==1.0.14
django-embed-video==1.1.2
django-redis==4.8.0
facebook-sdk==2.0.0
feedparser==5.2.1
kombu==4.0.2
mysqlclient==1.3.10
oauthlib==2.0.2
pytz==2017.2
redis==2.10.5
requests==2.14.2
requests-oauthlib==0.8.0
six==1.10.0
TwitterAPI==2.4.5
vine==1.1.3
